Feat : Updates project structure
