+++ app/context/RoomId/types/RoomContext.ts
@@ -1,0 +1,6 @@
+type RoomContext = {
+  roomId: string;
+  updateRoomId: (roomId: string) => void;
+};
+
+export default RoomContext;
