+++ app/context/UserName/types/UserContext.ts
@@ -1,0 +1,6 @@
+type UserContext = {
+  userName: string;
+  updateUserName: (userName: string) => void;
+};
+
+export default UserContext;
