+++ app/features/videoPlayback/types/FullScreenToggleProps.ts
@@ -1,0 +1,7 @@
+import { RefObject } from "react";
+
+type FullScreenToggleProps = {
+  videoWrapperRef: RefObject<HTMLDivElement>;
+};
+
+export default FullScreenToggleProps;
