+++ app/features/videoPlayback/types/PausePlayControlProps.ts
@@ -1,0 +1,7 @@
+type PausePlayControlProps = {
+  videoRef: React.RefObject<HTMLVideoElement>;
+  onManualPause?: (e: React.SyntheticEvent) => void;
+  onManualResume?: (e: React.SyntheticEvent) => void;
+};
+
+export default PausePlayControlProps;
