+++ app/features/videoPlayback/types/VideoMeta.ts
@@ -1,0 +1,6 @@
+type VideoMeta = {
+  duration: number;
+  currentTime: number;
+};
+
+export default VideoMeta;
