+++ app/features/videoPlayback/types/VideoPlayerButtonComponentProps.ts
@@ -1,0 +1,8 @@
+type VideoPlayerButtonComponentProps = {
+  children: React.ReactNode;
+  onClick?: (e: React.MouseEvent) => void;
+  onKeyPress?: (e: React.KeyboardEvent) => void;
+  allowSpacebarPress?: boolean;
+};
+
+export default VideoPlayerButtonComponentProps;
