+++ app/features/videoPlayback/types/VolumeControlProps.ts
@@ -1,0 +1,5 @@
+type VolumeControlProps = {
+  videoRef: React.RefObject<HTMLVideoElement>;
+};
+
+export default VolumeControlProps;
