--- app/routes/UserNameContextProvider.tsx
@@ -1,16 +1,0 @@
-import UserNameContext from "~/routes/UserNameContext";
-import { useState } from "react";
-
-export default function UserNameContextProvider({ children }:any) {
-    const [context, setContext] = useState('anonymous user');
-
-    const updateUserName = (userName: string) => {
-        setContext(userName);
-    }
-
-    return (
-        <UserNameContext.Provider value={{ userName:context, updateUserName }}>
-            {children}
-        </UserNameContext.Provider>
-    );
-}
